requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-auth-ldap==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
idna==2.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.15.0
python-ldap==3.3.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
